Refactor index.js to render StoryblokComponent and load draft content
The changes introduced in this commit refactor the `index.js` file in the `pages` directory to use `StoryblokComponent` for rendering content. The `useStoryblokState` hook is utilized to fetch and save the content states, while `getStoryblokApi` is used to obtain the draft version of the content. Additionally, the method `getStaticProps` has been updated to use the draft version of the content and load the necessary configuration from Storyblok's CMS.