feat(contact): add API route for contact form, validate form and send email
Added a new handler function for the contact form which returns an appropriate response code based on the method and body of the request. The handler function validates the request body and sends an email to the admin and the user. The validateContactForm util function was also added to validate the form fields. The email is sent through the Sendgrid mail API. The commit also includes some 'todo' comments for future improvements.