Refactor README layout for Portfolio Project
Changed original Readme layout into format that is more informative and allows for a clear overview of the project, including features, acknowledgements, and contact details. The project is a web app built with Next.js, @MUI/material, Twilio Sendgrid and StoryblokCMS and has a contact form with backend verification and email alerts.